JS-1394 Refactor S6767 false-positive escape helpers
